Overwrite old Infiniium driver. We are largely backwards compatible
